refactor(Rust): Create `parser-ts` crate
