+++ rust/parser-ts/Cargo.toml
@@ -1,0 +1,14 @@
+[package]
+name = "parser-ts"
+version = "0.0.0"
+edition = "2021"
+
+[dependencies]
+once_cell = "1.8.0"
+parser-treesitter = { path = "../parser-treesitter", version = "0.0.0" }
+parser-js = { path = "../parser-js", version = "0.0.0" }
+tree-sitter-typescript = "0.19.0"
+
+[dev-dependencies]
+test-snaps = { path = "../test-snaps", version = "0.0.0" }
+test-utils = { path = "../test-utils", version = "0.0.0" }
